msg/async: move ProtocolV1 class to its own source file
