+++ src/msg/CMakeLists.txt
@@ -21,7 +21,7 @@ endif(HAVE_XIO)
 list(APPEND msg_srcs
   async/AsyncConnection.cc
   async/AsyncMessenger.cc
-  async/Protocol.cc
+  async/ProtocolV1.cc
   async/Event.cc
   async/EventSelect.cc
   async/PosixStack.cc
